APIとSocket.io通信を別ファイルに切り分け
